weaviate_service/requirements.txt
fastapi==0.97.0
hazm==0.9.0
pydantic==1.9.2
python-dotenv==1.0.0
uvicorn==0.22.0
weaviate-client